# go/model_appointment_response.go
/*
 * Appointment Scheduling Service- OpenAPI 3.0
 *
 * The specification contains API contractrs for appointment scheduling service
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type AppointmentResponse struct {
	// The customer identifier that contains name, phone number and history
	CustomerID *interface{} `json:"customerID,omitempty"`
	// The starting time of the appointment.
	AppointmentTime string `json:"appointmentTime,omitempty"`
	// The item for the appointment.
	AppointmentItem string `json:"appointmentItem,omitempty"`
	// The duration of the appointment, this is associated with the item, and it can  be customed. (in minutes)
	AppointmentDuration int32 `json:"appointmentDuration,omitempty"`
}