# go/model_appointment_update_request.go
/*
 * Appointment Scheduling Service- OpenAPI 3.0
 *
 * The specification contains API contractrs for appointment scheduling service
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type AppointmentUpdateRequest struct {
	// The name of the customer, required if ID not provided
	CustomerName string `json:"customerName,omitempty"`
	// The phone number, required if the ID not provided
	CustomerNumber string `json:"customerNumber,omitempty"`
	// The starting time of the appointment.
	AppointmentTime string `json:"appointmentTime,omitempty"`
	// The item for the appointment.
	AppointmentItem string `json:"appointmentItem,omitempty"`
	// The duration of the appointment, this is associated with the item, and it can  be customed. (in minutes)
	AppointmentDuration int32 `json:"appointmentDuration,omitempty"`
}