# go/model_customer_update_request.go
/*
 * Appointment Scheduling Service- OpenAPI 3.0
 *
 * The specification contains API contractrs for appointment scheduling service
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type CustomerUpdateRequest struct {
	// The name of the customer, required if ID not provided
	CustomerName string `json:"customerName,omitempty"`
	// The phone number, required if the ID not provided
	CustomerNumber float64 `json:"customerNumber,omitempty"`
	// Remarks for the customer, e.g. discounts/ requests
	CustomerNote string `json:"customerNote,omitempty"`
	// The visit history of the customer
	AppointmentHistory *interface{} `json:"appointmentHistory,omitempty"`
}