# go/model_item_response.go
/*
 * Appointment Scheduling Service- OpenAPI 3.0
 *
 * The specification contains API contractrs for appointment scheduling service
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type ItemResponse struct {
	// The name of the item
	ItemName string `json:"itemName,omitempty"`
	// The description of the item
	ItemNotes string `json:"itemNotes,omitempty"`
	// the default price of the item, could be customised depending on situation
	ItemDefaultPrice float64 `json:"itemDefaultPrice,omitempty"`
	// The default duration of the item (in minutes), could be customised depending on situation
	ItemDefaultDuration int32 `json:"itemDefaultDuration,omitempty"`
}